add add button

+++ TodoList/Views/ListView.swift
@@ -9,23 +9,28 @@ import SwiftUI
 
 struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
-
     
     var body: some View {
         // 列表
-        List {
-            // for循环 类似 v-for
-            ForEach(listViewModel.items) { item in
-                // 对子view的引用
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                Text("No items")
+            } else {
+                List {
+                    // for循环 类似 v-for
+                    ForEach(listViewModel.items) { item in
+                        // 对子view的引用
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
         // 标题
         .navigationTitle("Todo List 📋")
@@ -36,13 +41,14 @@ struct ListView: View {
             // 尾
             trailing:
                 // 跳转到新的页面
-                NavigationLink(
-                    // 展示内容
-                    "Add",
-                    // 实际的跳转内容
-                    destination: AddView()
-                )
+            NavigationLink(
+                // 展示内容
+                "Add",
+                // 实际的跳转内容
+                destination: AddView()
+            )
         )
+        
     }
 }
